requirements.txt
# Core dependencies
pyyaml>=6.0
requests>=2.31.0

# AI Models
gpt4all>=2.5.0  # For GPT4All local LLM (no server needed)
# ollama  # Install separately: https://ollama.ai (if using Ollama)

# Telemetry (Windows shared memory)
pywin32>=306; sys_platform == "win32"

# Voice input/output
keyboard>=0.13.5
pyaudio>=0.2.14
openai-whisper>=20231117
pyttsx3>=2.90

# Optional: For better audio processing
numpy>=1.24.0
torch>=2.0.0; platform_system != "Darwin"  # torch can be heavy, optional for Mac

